Refactor: split integrations/groundcover/helpers.py by role (closes #5441)
